split delta test into three files for reading simplicity
